+++ R/itis_search_df.r
@@ -1,7 +1,7 @@
 #' Wrapper function for applying genus_search_itis and species_search_itis
 #' to a whole data.frame containing scientific names
 #' @param df data.frame containing names to check
-#' @param namecol integer or character string with column name containing 
+#' @param namecol integer or character string with column name containing
 #' 		  species or genus names
 #' @param higher Boolean. If TRUE, add higher taxonomic classifications to output
 #' @param genus.only boolean If TRUE, search for matches with just the genus name using genus_search_itis
@@ -11,21 +11,23 @@
 #' @return data.frame with submitted names (orig.name), matched names (matched.name),
 #' 1/0 flag indicating that original name is currently accepted (orig.name.accepted),
 #' 1/0 flag indicating if search was genus_only (for distinguishing genus_search_itis
-#' and species_search_itis results), synonyms if any, and higher taxonomy (if 
+#' and species_search_itis results), synonyms if any, and higher taxonomy (if
 #' higher=TRUE)
 #'
 #' @examples
-#  
+#
 #' data(lakegeneva)
-#' #example dataset with 50 rows
+#' #example dataset
 #'
-#' new.lakegeneva <- genus_species_extract(lakegeneva,'phyto_name')[1:5,]
-#'
+#' new.lakegeneva <- genus_species_extract(lakegeneva[1,],'phyto_name')
+#' new.lakegeneva$genus_species <- trimws(paste(new.lakegeneva$genus,
+#' new.lakegeneva$species))
 #' #checking for genus-only name matches in ITIS, and extracting higher taxonomy
 #' #flagging names with imperfect or no matches
 #'
-#' lakegeneva.genus.itischeck <- itis_search_df(new.lakegeneva,"genus",genus.only=TRUE)
-#' lakegeneva.genus.itischeck
+#' lakegeneva.genus.itischeck <-
+#'          itis_search_df(new.lakegeneva,"genus_species")
+#'          lakegeneva.genus.itischeck
 
 itis_search_df<-function(df,namecol=NA,higher=FALSE,genus.only=FALSE)
 {
@@ -36,12 +38,29 @@ itis_search_df<-function(df,namecol=NA,higher=FALSE,genus.only=FALSE)
 			itis.df<-rbind(itis.df,data.frame(Kingdom=NULL, Subkingdom=NULL, Infrakingdom=NULL, Phylum=NULL, Class=NULL,Subclass=NULL,Order=NULL,
 			Family=NULL)
 		)
+		}
+
+	#first, make sure itis is up.
+	itis.check<-RCurl::url.exists("www.itis.gov")
+	if(!itis.check){
+	  #if site is down, return warning
+	  message("Warning: Could not connect to ITIS website (www.itis.gov). The site may be
+            temporarily down. Please try again later.")
+	  #and populate return data.frame with NA.
+	  tmp<-data.frame(matched.name=NA,match=0,orig.name.accepted=0,
+	                  orig.name=df[[namecol]],genus.only=ifelse(genus.only,1,0),synonyms="")
+	  if(higher){
+	    tmp<-cbind(tmp,data.frame(Kingdom=NA, Subkingdom=NA, Infrakingdom=NA, Phylum=NA, Class=NA,Subclass=NA,Order=NA,
+	                              Family=NA)
+	    )
+	  }
+	  return(tmp)
 	}
 	#wrapper for genspp_itis_search and genus_itis_search on whole data.frame
-	#
+
 	nrows<-nrow(df)
 	for(i in 1:nrows)
-	{	
+	{
 		name=df[[namecol]][i]
 		if(!is.na(name) & !is.null(name))
 		{
@@ -50,7 +69,7 @@ itis_search_df<-function(df,namecol=NA,higher=FALSE,genus.only=FALSE)
 			}else{
 				tmp<-species_search_itis(name,higher=higher)
 			}
-			
+
 		}else{
 			tmp<-data.frame(matched.name=NA,match=0,orig.name.accepted=0,
 			orig.name=NA,genus.only=ifelse(genus.only,1,0),synonyms="")